рефакторинг MVC -> MVP - 1
