--- MovieQuizTests/ArrayTests.swift
@@ -1,30 +1,0 @@
-//
-//  ArrayTests.swift
-//  MovieQuizTests
-//
-//  Created by Ilya Lotnik on 22.04.2024.
-//
-
-import XCTest
-
-@testable import MovieQuiz
-
-class ArrayTests: XCTestCase {
-    func testGetValueInRange() throws {
-        let array = [1, 1, 2, 3, 5]
-        
-        let value = array[safe: 2]
-        
-        XCTAssertNotNil(value)
-        XCTAssertEqual(value, 2)
-        
-    }
-    
-    func testGetValueOutOfRange() throws {
-        let array = [1, 1, 2, 3, 5]
-        
-        let value = array[safe: 5]
-        
-        XCTAssertNil(value)
-    }
-}
